refactor(Session): Session, SessionManager를 catalina로 이동
