+++ tomcat/src/main/java/org/apache/catalina/SessionManager.java
@@ -1,7 +1,4 @@
-package org.apache.coyote.http11;
-
-import org.apache.catalina.Manager;
-import org.apache.catalina.Session;
+package org.apache.catalina;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
